% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopping.R
\name{build_skill_patterns}
\alias{build_skill_patterns}
\title{Build skill patterns matrix}
\usage{
build_skill_patterns(K)
}
\arguments{
\item{K}{Integer. Number of attributes.}
}
\value{
A matrix of 2^K rows and K columns.
}
\description{
Generates all 2^K binary skill profiles for K attributes.
}
