% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{.validate_initial_profile}
\alias{.validate_initial_profile}
\title{Validate initial_profile (internal)}
\usage{
.validate_initial_profile(initial_profile, K)
}
\arguments{
\item{initial_profile}{Binary integer vector of length K, or \code{NULL}.}

\item{K}{Integer. Number of attributes.}
}
\value{
Invisibly \code{TRUE}. Stops on error.
}
\description{
Validate initial_profile (internal)
}
\keyword{internal}
