% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{KL_criteria}
\alias{KL_criteria}
\title{KL divergence criterion}
\usage{
KL_criteria(item_index, alpha_hat_index, prob_matrix)
}
\arguments{
\item{item_index}{Integer. Index of the candidate item.}

\item{alpha_hat_index}{Integer. Index of the currently estimated profile.}

\item{prob_matrix}{Matrix J x 2^K of response probabilities.}
}
\value{
Numeric scalar.
}
\description{
Kullback-Leibler divergence between the estimated profile and all others.
}
