% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{MPWKL_criteria}
\alias{MPWKL_criteria}
\title{MPWKL criterion}
\usage{
MPWKL_criteria(item_index, prob_matrix, posterior)
}
\arguments{
\item{item_index}{Integer. Index of the candidate item.}

\item{prob_matrix}{Matrix J x 2^K of response probabilities.}

\item{posterior}{Numeric vector of posterior probabilities over profiles.}
}
\value{
Numeric scalar.
}
\description{
Mutual posterior-weighted Kullback-Leibler divergence.
}
