% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{PWKL_criteria}
\alias{PWKL_criteria}
\title{PWKL criterion}
\usage{
PWKL_criteria(item_index, alpha_hat_index, prob_matrix, posterior)
}
\arguments{
\item{item_index}{Integer. Index of the candidate item.}

\item{alpha_hat_index}{Integer. Index of the currently estimated profile.}

\item{prob_matrix}{Matrix J x 2^K of response probabilities.}

\item{posterior}{Numeric vector of posterior probabilities over profiles.}
}
\value{
Numeric scalar.
}
\description{
Posterior-weighted Kullback-Leibler divergence.
}
